+++ CotEditor/Sources/MultipleReplacementPanelController.swift
@@ -8,7 +8,7 @@
 //
 //  ---------------------------------------------------------------------------
 //
-//  © 2018 1024jp
+//  © 2018-2020 1024jp
 //
 //  Licensed under the Apache License, Version 2.0 (the "License");
 //  you may not use this file except in compliance with the License.
@@ -34,6 +34,10 @@ final class MultipleReplacementPanelController: NSWindowController {
         super.windowDidLoad()
         
         self.windowFrameAutosaveName = NSWindow.FrameAutosaveName("Multiple Replacement Panel")
+        
+        if ProcessInfo().operatingSystemVersion.majorVersion < 11 {
+            self.window?.styleMask.remove(.fullSizeContentView)
+        }
     }
     
 }